Create README.md
The estimation of food calories from images has become an essential research area due to 
the increasing interest in health-consciousness and dietary monitoring. In this paper, the 
CNN algorithm is combined with deep learning. Here, the dataset is made up of the 
photos. Those pictures underwent feature-based training. These pictures were used as the  
input, and it provides the image’s name and the number of calories. The CNN algorithm 
has been used in this process. Once the food items are identified, a food recognition model 
based on deep learning is applied to classify the detect the items. To evaluate the 
performance of the proposed system, extensive experiments are conducted on a diverse 
dataset of food images. The application of this system extends to various domains, 
including diet monitoring applications, restaurant menu analysis, and dietary planning for 
individuals. By providing users with a convenient and reliable way to estimate food 
calories from images, the system can play a significant role in promoting healthier eating 
habits and managing calorie intake effectively. Obesity and Overweight is a common and 
serious disease in most of the adults and children, which refers to abnormal or excessive 
accumulation of fat that causes serious health problems. This paper presents a novel 
approach to food recognition using deep learning techniques, leveraging convolutional 
neural networks (CNNs) to accurately identify a wide variety of food items from images.